Extract functions to diary-food-table
Extracts getFoods and makeFoodTable function to diary-food-table.js.
Modifies diary.js to call appropriate function from exports on
document ready. All functionality is as prior to refactor.
