+++ lib/diary/diary.js
@@ -1,5 +1,6 @@
 const API = 'https://qs-be.herokuapp.com/api/v1/'
 const LOCAL = 'http://localhost:3000/api/v1/'
+const DiaryFoodTable = require('./diary-food-table')
 $ = require('jquery')
 
 const mealNames = ["Breakfast", "Lunch", "Dinner", "Snacks"]
@@ -33,20 +34,6 @@ const Diary = class Diary {
     .fail( (error) => {
       console.error(error)
     })
-  }
-
-
-  getFoods() {
-    return $.ajax({
-    method: 'GET',
-    url: API + 'foods',
-    })
-    .then((data) => {
-      makeFoodTable(data)
-    })
-    .fail( (error) => {
-      console.error(error)
-    });
   }
 
   addMeal() {
@@ -164,12 +151,6 @@ function resetCalories() {
   $('#remaining-calories').css('color', 'green').text(2000)
 }
 
-function makeFoodTable(data) {
-  data.forEach( (food) => {
-    $('#diary-foods-body').prepend(`<tr data-food-id=${food.id}><td class='food-cell' >${food.name}</td><td class='calorie-cell'>${food.calories}</td><td><input type="checkbox" id="${food.id}" data-food-name="${food.name}"/><label for="${food.id}"></label></td></tr>`)
-  })
-}
-
 function updateMealRemainingCalories(mealName, food){
   let currentValue = parseInt($(`#${mealName}-remaining-calories`).text())
   let updatedValue = currentValue - food.calories
@@ -179,7 +160,6 @@ function updateMealRemainingCalories(mealName, food){
   } else if (updatedValue >= 0) {
     $(`#${mealName}-remaining-calories`).css('color', 'green')
   }
-
 }
 
 function addFoodstoMealTable(meal, mealName){
@@ -296,9 +276,7 @@ function removeFoodFromMeal(){
     url: `${API}meals/${id}`,
     method: 'DELETE'
   })
-  .then((button) => {
-    removeFoodFromMealTable(button)
-  })
+  .then(removeFoodFromMealTable(button))
   .fail((error) => {
     console.error(error)
   })
@@ -310,7 +288,7 @@ $(document).ready(function(){
     searchFoods()
   })
 
-  diary.getFoods()
+  DiaryFoodTable.getFoods()
 
   diary.getDiary()
