+++ lib/index.js
@@ -1,3 +1,3 @@
 // const Food =
 require('./food/food');
-require('./diary')
+require('./diary/diary')
